+++ docs/source/changelog.rst
@@ -33,6 +33,7 @@ Ksconf v0.13.5 (DRAFT)
    Hopefully not many people are relying on this function yet, at it was only introduced 2 moths ago in v0.11.5.
    This function will be removed in v0.20.0.
 *  Created new :py:func:`~ksconf.conf.parser.write_conf_string` function for writing a conf file to a python string.
+*  Add new ``--empty-stanza`` to :ref:`ksconf_cmd_filter`.  Thanks to Manoj Baba for the idea.
 
 
 Ksconf v0.13.5 (2023-10-20)